- Add WebViewModel.swift
